+++ config-overrides.js
@@ -1,0 +1,4 @@
+module.exports = function override(config, env) {
+    return config;
+  };
+
