--- src/components/ParticleBG/ParticleBG.module.css
@@ -1,8 +1,0 @@
-/* src/components/ParticlesBackground/ParticlesBackground.module.css */
-.particlesBackground {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    z-index: -1; 
-  }
-
