+++ src/components/ParticlesBackground/ParticlesBackground.module.css
@@ -1,0 +1,7 @@
+.particlesContainer {
+  position: absolute; 
+  width: 100%; 
+  height: 100%; 
+  z-index: -1; 
+  overflow: hidden;
+}
